+++ src/dsp_tools/commands/xmlupload/xmlupload.py
@@ -15,6 +15,7 @@ from dsp_tools.clients.legal_info_client import LegalInfoClient
 from dsp_tools.clients.legal_info_client_live import LegalInfoClientLive
 from dsp_tools.clients.list_client import ListGetClient
 from dsp_tools.clients.list_client_live import ListGetClientLive
+from dsp_tools.clients.project_client_live import ProjectClientLive
 from dsp_tools.commands.validate_data.validate_data import validate_parsed_resources
 from dsp_tools.commands.xmlupload.execute_upload import enable_unknown_license_if_any_are_missing
 from dsp_tools.commands.xmlupload.execute_upload import execute_upload
@@ -64,6 +65,10 @@ def xmlupload(
     shortcode = root.attrib["shortcode"]
 
     auth = AuthenticationClientLive(server=creds.server, email=creds.user, password=creds.password)
+
+    # verify that a project with the shortcode exists
+    ProjectClientLive(creds.server, auth).get_project_iri(shortcode)
+
     config = config.with_server_info(server=creds.server, shortcode=shortcode)
     clients = _get_live_clients(auth, creds, shortcode, imgdir)